+++ development/web/lib/development_web/live/showcase/slider_color_form_demo.ex
@@ -32,6 +32,7 @@ defmodule DevelopmentWeb.Showcase.SliderColorFormDemo do
     ~H"""
     <div class="max-w-xs">
       <.form
+        id={"#{@id}-form"}
         for={to_form(%{"v" => @value}, as: :slider_demo)}
         phx-target={@myself}
         phx-submit="save"